chore: move some timeline code from corelib to job-worker
